requirements.txt
azureml-core
numpy
pandas
tqdm
fastparquet
pyarrow
